+++ wazo_auth/flask_helpers.py
@@ -20,23 +20,8 @@ class Tenant:
         self.name = name
 
     @classmethod
-    def autodetect(cls, many=False):
+    def autodetect(cls):
         specified_tenant = request.headers.get('Wazo-Tenant')
-
-        if not many:
-            return cls._one(specified_tenant)
-
-        tenants = cls._autodetect()
-
-        authorized_tenants = [t.uuid for t in tenants if t.uuid == specified_tenant]
-        if specified_tenant and specified_tenant not in authorized_tenants:
-            logger.debug('specified tenant not in available tenants')
-            raise tenant_helpers.UnauthorizedTenant(specified_tenant)
-
-        return cls._many(tenants, specified_tenant)
-
-    @classmethod
-    def _one(cls, specified_tenant):
         token = cls._get_token()
         user_uuid = token.metadata.get('uuid')
         logger.debug('token.metadata %s', token.metadata)
@@ -60,21 +45,6 @@ class Tenant:
         raise tenant_helpers.UnauthorizedTenant(specified_tenant)
 
     @classmethod
-    def _many(cls, tenants, specified_tenant):
-        if specified_tenant:
-            tenants = [t for t in tenants if t.uuid == specified_tenant]
-        return tenants
-
-    @classmethod
-    def _autodetect(cls):
-        token = cls._get_token()
-        user_uuid = token.metadata.get('uuid')
-        if user_uuid:
-            return cls._get_user_tenants(user_uuid)
-
-        return [cls(t['uuid'], t['name']) for t in token.metadata['tenants']]
-
-    @classmethod
     def _get_token(cls):
         token_uuid = request.headers.get('X-Auth-Token')
         return cls._get_token_data(token_uuid)
@@ -89,11 +59,6 @@ class Tenant:
         return user['tenant_uuid']
 
     @classmethod
-    def _get_user_tenants(cls, user_uuid):
-        user_tenants = cls.user_service.list_tenants(user_uuid)
-        return [cls(t['uuid'], t['name']) for t in user_tenants]
-
-    @classmethod
     def _get_all_sub_tenants(cls, tenant_uuid):
         return cls.tenant_service.list_sub_tenants(tenant_uuid)
 
@@ -102,6 +67,9 @@ class Tenant:
         cls.token_service = token_service
         cls.user_service = user_service
         cls.tenant_service = tenant_service
+
+    def visible_tenants(self):
+        return self.tenant_service.list_sub_tenants(self.uuid)
 
 
 class Token:
@@ -116,3 +84,10 @@ class Token:
     def from_headers(cls):
         token_uuid = request.headers.get('X-Auth-Token')
         return cls.token_service.get(token_uuid, required_access=None)
+
+
+def get_tenant_uuids(recurse=False):
+    tenant = Tenant.autodetect()
+    if not recurse:
+        return [tenant.uuid]
+    return tenant.visible_tenants()