+++ wazo_auth/plugins/http/config/plugin.py
@@ -11,5 +11,7 @@ class Plugin:
         config = dependencies['config']
         config_service = ConfigService(config)
         api.add_resource(
-            ConfigResource, '/config', resource_class_args=[config_service]
+            ConfigResource,
+            '/config',
+            resource_class_args=[config_service],
         )