+++ wazo_auth/plugins/http/group_policy/plugin.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from . import http
@@ -10,11 +10,15 @@ class Plugin:
         args = (dependencies['group_service'], dependencies['policy_service'])
 
         api.add_resource(
-            http.GroupPolicy,
+            http.GroupPolicyUUID,
             '/groups/<uuid:group_uuid>/policies/<uuid:policy_uuid>',
             resource_class_args=args,
         )
-
+        api.add_resource(
+            http.GroupPolicySlug,
+            '/groups/<uuid:group_uuid>/policies/<string:policy_slug>',
+            resource_class_args=args,
+        )
         api.add_resource(
             http.GroupPolicies,
             '/groups/<uuid:group_uuid>/policies',