+++ wazo_auth/plugins/http/groups/plugin.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from . import http
@@ -10,6 +10,12 @@ class Plugin:
         args = (dependencies['group_service'],)
 
         api.add_resource(
-            http.Group, '/groups/<uuid:group_uuid>', resource_class_args=args
+            http.Group,
+            '/groups/<uuid:group_uuid>',
+            resource_class_args=args,
         )
-        api.add_resource(http.Groups, '/groups', resource_class_args=args)
+        api.add_resource(
+            http.Groups,
+            '/groups',
+            resource_class_args=args,
+        )