+++ wazo_auth/plugins/http/password_reset/plugin.py
@@ -1,4 +1,4 @@
-# Copyright 2018-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2018-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from wazo_auth.http import AuthClientFacade
@@ -15,5 +15,7 @@ class Plugin:
         )
 
         api.add_resource(
-            http.PasswordReset, '/users/password/reset', resource_class_args=args
+            http.PasswordReset,
+            '/users/password/reset',
+            resource_class_args=args,
         )