+++ wazo_auth/plugins/http/status/plugin.py
@@ -1,4 +1,4 @@
-# Copyright 2020 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2020-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from xivo.status import Status
@@ -13,7 +13,11 @@ class Plugin:
 
         status_aggregator.add_provider(provide_status)
 
-        api.add_resource(StatusList, '/status', resource_class_args=[status_aggregator])
+        api.add_resource(
+            StatusList,
+            '/status',
+            resource_class_args=[status_aggregator],
+        )
 
 
 def provide_status(status):