+++ wazo_auth/plugins/http/tenants/plugin.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from . import http
@@ -9,7 +9,13 @@ class Plugin:
         api = dependencies['api']
         args = (dependencies['tenant_service'],)
 
-        api.add_resource(http.Tenants, '/tenants', resource_class_args=args)
         api.add_resource(
-            http.Tenant, '/tenants/<string:tenant_uuid>', resource_class_args=args
+            http.Tenants,
+            '/tenants',
+            resource_class_args=args,
+        )
+        api.add_resource(
+            http.Tenant,
+            '/tenants/<string:tenant_uuid>',
+            resource_class_args=args,
         )