+++ wazo_auth/plugins/http/user_group/plugin.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from . import http
@@ -15,9 +15,10 @@ class Plugin:
             resource_class_args=args,
         )
         api.add_resource(
-            http.GroupUsers, '/groups/<uuid:group_uuid>/users', resource_class_args=args
+            http.GroupUsers,
+            '/groups/<uuid:group_uuid>/users',
+            resource_class_args=args,
         )
-
         api.add_resource(
             http.UserGroups,
             '/users/<uuid:user_uuid>/groups',