+++ wazo_auth/plugins/http/user_registration/plugin.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from . import http
@@ -13,4 +13,8 @@ class Plugin:
             dependencies['user_service'],
         )
 
-        api.add_resource(http.Register, '/users/register', resource_class_args=args)
+        api.add_resource(
+            http.Register,
+            '/users/register',
+            resource_class_args=args,
+        )